requirements.txt
fastapi
uvicorn
python-multipart
requests
lmnt
python-dotenv
numpy
pillow